Refactoring to align with main
