+++ app/src/main/java/com/CMPUT301F21T19/habitappt/Fragments/AllHabits.java
@@ -16,13 +16,12 @@ import androidx.annotation.Nullable;
 import java.util.ArrayList;
 
 import com.CMPUT301F21T19.habitappt.Entities.Habit;
-import com.CMPUT301F21T19.habitappt.recycler_view_fragment;
 import com.google.firebase.firestore.FirebaseFirestoreException;
 import com.google.firebase.firestore.QueryDocumentSnapshot;
 import com.google.firebase.firestore.QuerySnapshot;
 
 
-public class AllHabits extends recycler_view_fragment {
+public class AllHabits extends RecyclerViewFragment {
     /**
      * This class defines a fragment that shows all of a users habits.
      */
